A Pinch of Improvement
A little bit goes a long way! I finished painting the walls and floor, straightened out some roofs, and did a couple of other minor adjustments to fix a few anomalies in the level. I also added a roof to the level as a whole. Oh, and...

I found a way to shrink the level! If this causes any problems that cannot easily be fixed, this is very easy to undo: just click on the "everything" object (which is a parent for the entire level) and change its scale factor from 0.1 to 1 on all axes.